Well, I was able to create a concert so far, whether it is correct is a matter for another commit xd

+++ handler/section.go
@@ -13,7 +13,7 @@ import (
 
 type CreateSection struct {
 	Name         string `json:"name"`
-	InstrumentID int64  `json:"instrumentId"`
+	InstrumentID uint   `json:"instrumentId"`
 }
 
 type SectionHandler struct {
